% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camel.R
\name{camel}
\alias{camel}
\title{camel}
\usage{
camel(x)
}
\arguments{
\item{x}{}
}
\value{
text converted to camel case
}
\description{
converts text to camel case.
}
\examples{
camel("hola mundo") -> "holaMundo"
}

