% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewRealEstateData.R
\name{getNewRealEstateData}
\alias{getNewRealEstateData}
\title{TitlegetNewRealEstateData}
\usage{
getNewRealEstateData(database = "RealEstate.csv",
  backup = "RealEstateBackup.csv")
}
\arguments{
\item{database}{}

\item{backup}{}
}
\value{
NA
}
\description{
Scraps the sources to extract new real estate data. Adds new records to the database. creates a backip of the previous version of the database.
}
\examples{
getNewRealEstateData()
}

