% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryPriceByM2ConstruccionByProvincia.R
\name{getSummaryPriceByM2ConstruccionByProvincia}
\alias{getSummaryPriceByM2ConstruccionByProvincia}
\title{getSummaryPriceByM2ConstruccionByProvincia}
\usage{
getSummaryPriceByM2ConstruccionByProvincia(data)
}
\arguments{
\item{data}{}
}
\description{
getSummaryPriceByM2ConstruccionByProvincia
}

