% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryPriceInUSDByProvincia.R
\name{getSummaryPriceInUSDByProvincia}
\alias{getSummaryPriceInUSDByProvincia}
\title{getSummaryPriceInUSDByProvincia}
\usage{
getSummaryPriceInUSDByProvincia(data)
}
\arguments{
\item{data}{}
}
\description{
getSummaryPriceInUSDByProvincia
}

