% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareForAnalysis.R
\name{prepareForAnalysis}
\alias{prepareForAnalysis}
\title{prepareForAnalysis}
\usage{
prepareForAnalysis(data)
}
\arguments{
\item{data}{}
}
\description{
creates new columns useful to analyze data
}

